refactor: move retrieve_blocking into Shard crate
