+++ lib/shard/src/retrieve/mod.rs
@@ -1,1 +1,2 @@
 pub mod record_internal;
+pub mod retrieve_blocking;
